requirements.txt
appdirs==1.4.3
appnope==0.1.0
decorator==4.0.6
gnureadline==6.3.3
ipython-genutils==0.2.0
numpy==1.12.1
packaging==16.8
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
ptyprocess==0.5
pyparsing==2.2.0
redis==2.10.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
cv2
